refactor: Nest cost-services tools.
