+++ src/tools/cost-services/index.ts
@@ -1,0 +1,1 @@
+import "./list-cost-services";
