+++ src/tools/index.ts
@@ -8,6 +8,7 @@ import "./business-metrics";
 import "./canvases";
 import "./cost-alerts";
 import "./cost-reports";
+import "./cost-services";
 import "./costs";
 import "./create-virtual-tag-config";
 import "./current-user";
@@ -17,7 +18,6 @@ import "./folders";
 import "./get-cost-provider-accounts";
 import "./list-cost-integrations";
 import "./list-cost-providers";
-import "./list-cost-services";
 import "./network-flow-reports";
 import "./provider-resources";
 import "./recommendation-views";
